Optimize event data fetching
The JOIN statements generated by sequelize copies over the data from
each selected column to every row, leading to a database query with a
massive amount of redundant data. By separating these queries the data
passed between the database and node is cut down significantly.

Additionally, this separation of event and quota querying means that
only answers to public questions need to be queried, further reducing
the size of the returned data greatly.